Move core to lib/draught
We're extracting the core to a gem, so the first step is to move
everything across to the new folder structure and ensure the tests
pass
