+++ lib/draught/bounding_box.rb
