+++ lib/draught/boxlike.rb
