+++ lib/draught/corner.rb
