+++ lib/draught/cubic_bezier.rb
