+++ lib/draught/curve.rb
