+++ lib/draught/line.rb
