+++ lib/draught/path.rb
