+++ lib/draught/path_builder.rb
