+++ lib/draught/path_cleaner.rb
