+++ lib/draught/pathlike.rb
