+++ lib/draught/point.rb
