+++ lib/draught/pointlike.rb
