+++ lib/draught/renderer.rb
