+++ lib/draught/transformations.rb
