+++ lib/draught/transformations/affine.rb
