+++ lib/draught/transformations/common.rb
