+++ lib/draught/transformations/composer.rb
