+++ lib/draught/transformations/composition.rb
