+++ lib/draught/transformations/proclike.rb
