+++ lib/draught/vector.rb
