+++ spec/draught/pathlike_examples.rb
