+++ spec/draught/transformations/proclike_spec.rb
@@ -1,6 +1,6 @@
-require 'box_dividers/transformations/proclike'
-require 'box_dividers/transformations/shared_examples'
-require 'box_dividers/point'
+require 'draught/transformations/proclike'
+require 'draught/transformations/shared_examples'
+require 'draught/point'
 
 module BoxDividers::Transformations
   RSpec.describe Proclike do
